requirements.txt
kafka-python==2.0.2
requests==2.27.1
schedule==1.1.0

flake8==4.0.1
black==22.3.0
